Build .Jar
A bug that caused the .jar to not be correctly executed due to the "setSongAvatar" method being repeated endlessly was fixed.

Also, "setSongAvatar" was overloaded for the purpose of avoiding workarounds, such as passing a null image on the original method.

The function "createImage" was added to simplify the instantiation of image files.